Modify prim_fifo_async to prevent data leakage
This change modifies prim_fifo_async to split its first entry into a
separate register which is initialized to zero on reset. This prevents
stale data in the memory from being readable on rdata before data is
written to the FIFO.

Although a more straightforward option would be to set the rdata output
to 0 if rvalid is 0, this breaks SPI TX functionality since it relies on
reading a byte from an async FIFO before the FIFO control signals can
propagate from the write to read clock domain.